Convert contribution owner to ID, use smaller number formats

+++ apps/token/contracts/Token.sol
@@ -9,15 +9,15 @@ contract Token is ERC20Token, AragonApp {
   // ensure alphabetic order
   enum Apps { Contribution, Contributor, Proposal, Token }
   bytes32[4] public appIds;
- 
-  event LogMint(address indexed recipient, uint256 amount, uint256 contributionId);
-  
+
+  event LogMint(address indexed recipient, uint256 amount, uint32 contributionId);
+
   function initialize(bytes32[4] _appIds) public onlyInit {
     appIds = _appIds;
     initialized();
   }
 
-  function mintFor(address contributorAccount, uint256 amount, uint256 contributionId) public isInitialized auth(MINT_TOKEN_ROLE) {
+  function mintFor(address contributorAccount, uint256 amount, uint32 contributionId) public isInitialized auth(MINT_TOKEN_ROLE) {
     _mint(contributorAccount, amount);
     emit LogMint(contributorAccount, amount, contributionId);
   }
